+++ lib/converter/configurator/abstract_converter.dart
@@ -2,7 +2,7 @@ import 'package:flutter_quill_to_pdf/packages/html2md/lib/html2md.dart' as hm2;
 import 'package:flutter_quill_to_pdf/utils/makdown_rules_custom.dart';
 
 ///Converter is a parent that's [provides the essential rules]
-///for generate and create the [book] file for that project
+///for generate and create a document pdf file for that project
 sealed class Converter<Doc, Type> {
   late final List<hm2.Rule> _rules;
   late final Doc _document;
@@ -16,12 +16,14 @@ sealed class Converter<Doc, Type> {
   List<hm2.Rule> get rules => _rules;
   Doc get document => _document;
 
+  ///This fuction let us add new rules as we wants. We can clear the default rules, or just write our rules. 
+  ///You can check [html2md] documentation about here: https://github.com/jarontai/html2md
   void customRules(List<hm2.Rule> customRules, {bool clearDefaultRules = false}) {
     if (clearDefaultRules) _rules.clear();
     _rules.addAll(customRules);
   }
 
-  ///This functions creates the [type document] selected for us file using the widgets that are provides by [generatePages] functions
+  ///This functions generates the [document]
   Future<Type> generateDoc();
 }